chore(devnet-sdk): relocate telemetry package
